TestRender project removed.

+++ Raytracer/UnitTests/GridDensityVolumeRegion.test.h
@@ -26,9 +26,9 @@ class GridDensityVolumeRegionTestSuite : public CxxTest::TestSuite
       m_size_y=110;
       m_size_z=120;
       m_densities.assign(m_size_x, std::vector<std::vector<double> >(m_size_y, std::vector<double>(m_size_z, 0.0)));
-      for(int i=0;i<m_size_x;++i)
-        for(int j=0;j<m_size_y;++j)
-          for(int k=0;k<m_size_z;++k)
+      for(size_t i=0;i<m_size_x;++i)
+        for(size_t j=0;j<m_size_y;++j)
+          for(size_t k=0;k<m_size_z;++k)
             m_densities[i][j][k]=i+j+k;
 
       mp_volume.reset(new GridDensityVolumeRegion<PhaseFunctionMock>(m_bounds, m_emission, m_absorption, m_scattering, phase_function, m_densities));
